create_people module created
