+++ src/modules/user_choice.rb
@@ -1,6 +1,9 @@
 require_relative '../../app'
+# require_relative './list'
 
-def main
+# include List
+
+def menu
   puts "\n Welcome to School Library App! \n\n"
   puts "Please chose an option by entering a number: \n\n"
   options = {
@@ -13,30 +16,29 @@ def main
     7 => 'Exit'
   }
   puts options.map { |key, value| "#{key}. #{value}" }.join("\n")
+end
 
-  def option_case(choice)
-    case choice
-    when 1
-      load_books
-    when 2
-      load_people
-    when 3
-      create_person
-    when 4
-      create_book
-    when 5
-      create_rental
-    when 6
-       load_rentals
-    else
-      puts '*** Incorect choice, chose a number between 1-7 ***'
-    end
+def option_case(choice)
+  case choice
+  when 1
+    load_books
+  when 2
+    load_people
+  when 3
+    create_person
+  when 4
+    create_book
+  when 5
+    create_rental
+  when 6
+    load_rentals
+  else
+    puts '*** Incorect choice, chose a number between 1-7 ***'
   end
-  
-  def prompt_user
-    main
-    choice = gets.chomp.to_i
-    choice == 7 ? quit_app : option_case(choice)
-    # prompt_user
-  end
+end
+
+def prompt_user
+  menu
+  choice = gets.chomp.to_i
+  choice == 7 ? quit_app : option_case(choice)
 end
